add: contact/newのView・Controllerを追加
