+++ Gemfile
@@ -59,6 +59,9 @@ gem "mini_magick"
 # 2024/03/08
 gem "jquery-rails" # bootstrapを読み込むため
 
+# 2024/03/14
+gem "rails-i18n"
+
 group :development, :test do
   # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
   gem "debug", platforms: %i[ mri mingw x64_mingw ]
