+++ app/assets/stylesheets/contact.scss
@@ -1,0 +1,3 @@
+#contactEntrance {
+  width: fit-content;
+}
