+++ app/views/shared/_about_modal.erb
@@ -16,7 +16,7 @@
           <br>
           <li>学籍登録(後日実装予定)を行うと、<br>表示されるセクション名やチャイムのなる時刻を編集できるようになります。</li>
           <br>
-          <p class="text-end" style="font-size:12px; opacity:0.5;">チャイム音提供: OtoLogic(https://otologic.jp)</p>
+          <p class="text-end" id="chimeCredit">チャイム音提供: OtoLogic(https://otologic.jp)</p>
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
