+++ config/locales/activerecord/ja.yml
@@ -1,0 +1,7 @@
+ja:
+  activerecord:
+    attributes:
+      contact:
+        name: お名前
+        email: メールアドレス
+        content: お問い合わせ内容
