+++ config/routes.rb
@@ -3,7 +3,7 @@ Rails.application.routes.draw do
 
   resources :contacts, only: %i[new create] do
     collection do
-      get "check", to: "contacts#check"
+      post "check", to: "contacts#check"
       get "completed", to: "contacts#completed"
     end
   end
